+++ src/main/kotlin/joshxviii/plantz/effect/GardenHeroEffect.kt
@@ -1,13 +1,17 @@
 package joshxviii.plantz.effect
 
+import joshxviii.plantz.PazLootTables
+import net.minecraft.resources.ResourceKey
 import net.minecraft.server.level.ServerLevel
 import net.minecraft.world.effect.MobEffect
 import net.minecraft.world.effect.MobEffectCategory
 import net.minecraft.world.entity.LivingEntity
+import net.minecraft.world.level.storage.loot.LootTable
 
 class GardenHeroEffect(
     category: MobEffectCategory,
-    color: Int
+    color: Int,
+    var lootTables: List<ResourceKey<LootTable>> = listOf(PazLootTables.GARDEN_HERO_MAIL)
 ) : MobEffect(category, color) {
 
     override fun shouldApplyEffectTickThisTick(remainingDuration: Int, amplification: Int): Boolean {
@@ -17,6 +21,5 @@ class GardenHeroEffect(
     override fun applyEffectTick(level: ServerLevel, mob: LivingEntity, amplification: Int): Boolean {
         return true
     }
-
 
 }